# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Grapevine
  description: Course and advisor review platform for IIIT Hyderabad
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: local dev
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: id
  responses:
    Unauthorized:
      description: not logged in
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    Forbidden:
      description: insufficient permissions
      content:
        text/plain:
          schema:
            type: string
            example: forbidden
    NotFound:
      description: resource not found
      content:
        text/plain:
          schema:
            type: string
            example: not found
    BadRequest:
      description: bad request
      content:
        text/plain:
          schema:
            type: string
  schemas:
...
    Ulid:
      type: string
      description: ULID identifier
      example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
    Rating:
      type: integer
      minimum: 1
      maximum: 5
...
    AuthorRef:
      type: object
      required: [id, display_name]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        display_name:
          type: string

    # ── courses ─────────────────────────────────────────────────────────────────
...
    CourseReview:
      type: object
      required: [id, offering_id, anonymous, difficulty, teaching, grading, content, workload, overall, body, score, created_at]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        offering_id:
          $ref: '#/components/schemas/Ulid'
        author:
          allOf:
            - $ref: '#/components/schemas/AuthorRef'
          nullable: true
          description: null when the review is anonymous
        anonymous:
          type: boolean
        difficulty:
          $ref: '#/components/schemas/Rating'
        teaching:
          $ref: '#/components/schemas/Rating'
        grading:
          $ref: '#/components/schemas/Rating'
        content:
          $ref: '#/components/schemas/Rating'
        workload:
          $ref: '#/components/schemas/Rating'
        overall:
          type: number
          format: float
          description: mean of the five axes, rounded to 2 decimal places
          example: 3.8
        body:
          type: string
        score:
          type: integer
          description: sum of all votes (upvotes minus downvotes)
        user_vote:
          type: integer
          nullable: true
          description: "current user's vote: 1 (up), -1 (down), or null if no vote cast"
        edited_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    AdvisorReview:
      type: object
      required: [id, faculty_id, anonymous, research, availability, mentorship, support, workload, overall, body, score, created_at]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        faculty_id:
          $ref: '#/components/schemas/Ulid'
        author:
          allOf:
            - $ref: '#/components/schemas/AuthorRef'
          nullable: true
          description: null when the review is anonymous
        anonymous:
          type: boolean
        research:
          $ref: '#/components/schemas/Rating'
        availability:
          $ref: '#/components/schemas/Rating'
        mentorship:
          $ref: '#/components/schemas/Rating'
        support:
          $ref: '#/components/schemas/Rating'
        workload:
          $ref: '#/components/schemas/Rating'
        overall:
          type: number
          format: float
          description: mean of the five axes, rounded to 2 decimal places
          example: 4.2
        body:
          type: string
        score:
          type: integer
        user_vote:
          type: integer
          nullable: true
          description: "1 (up), -1 (down), or null if no vote cast"
        edited_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── review request bodies ────────────────────────────────────────────────────
    CreateCourseReview:
      type: object
      required: [anonymous, difficulty, teaching, grading, content, workload, body]
      properties:
        anonymous:
          type: boolean
        difficulty:
          $ref: '#/components/schemas/Rating'
        teaching:
          $ref: '#/components/schemas/Rating'
        grading:
          $ref: '#/components/schemas/Rating'
        content:
          $ref: '#/components/schemas/Rating'
        workload:
          $ref: '#/components/schemas/Rating'
        body:
          type: string
    EditCourseReview:
      type: object
      description: all fields optional; only provided fields are updated
      properties:
        difficulty:
          $ref: '#/components/schemas/Rating'
        teaching:
          $ref: '#/components/schemas/Rating'
        grading:
          $ref: '#/components/schemas/Rating'
        content:
          $ref: '#/components/schemas/Rating'
        workload:
          $ref: '#/components/schemas/Rating'
        body:
          type: string
    CreateAdvisorReview:
      type: object
      required: [anonymous, research, availability, mentorship, support, workload, body]
      properties:
        anonymous:
          type: boolean
        research:
          $ref: '#/components/schemas/Rating'
        availability:
          $ref: '#/components/schemas/Rating'
        mentorship:
          $ref: '#/components/schemas/Rating'
        support:
          $ref: '#/components/schemas/Rating'
        workload:
          $ref: '#/components/schemas/Rating'
        body:
          type: string
    EditAdvisorReview:
      type: object
      description: all fields optional; only provided fields are updated
      properties:
        research:
          $ref: '#/components/schemas/Rating'
        availability:
          $ref: '#/components/schemas/Rating'
        mentorship:
          $ref: '#/components/schemas/Rating'
        support:
          $ref: '#/components/schemas/Rating'
        workload:
          $ref: '#/components/schemas/Rating'
        body:
          type: string
...
paths:
...
  /offerings/{id}/reviews:
    get:
      operationId: listOfferingReviews
      summary: list reviews for a specific offering
      tags: [offerings, reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Ulid'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseReview'

    post:
      operationId: createCourseReview
      summary: submit a course review for this offering
      tags: [offerings, reviews]
      security:
        - session: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Ulid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseReview'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseReview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── faculty ─────────────────────────────────────────────────────────────────
...
  /faculty/{slug}/reviews:
    get:
      operationId: listAdvisorReviews
      summary: list advisor reviews for a faculty member
      tags: [faculty, reviews]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            example: girish-varma
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdvisorReview'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      operationId: createAdvisorReview
      summary: submit an advisor review for this faculty member
      tags: [faculty, reviews]
      security:
        - session: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            example: girish-varma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdvisorReview'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorReview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── labs ────────────────────────────────────────────────────────────────────
...
  /reviews/course/{id}:
    patch:
      operationId: editCourseReview
      summary: edit own course review
      tags: [reviews]
      security:
        - session: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Ulid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditCourseReview'
      responses:
        '200':
          description: updated review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseReview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      operationId: deleteCourseReview
      summary: delete own course review
      tags: [reviews]
      security:
        - session: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Ulid'
      responses:
        '204':
          description: deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /reviews/advisor/{id}:
    patch:
      operationId: editAdvisorReview
      summary: edit own advisor review
      tags: [reviews]
      security:
        - session: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Ulid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditAdvisorReview'
      responses:
        '200':
          description: updated review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorReview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      operationId: deleteAdvisorReview
      summary: delete own advisor review
      tags: [reviews]
      security:
        - session: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Ulid'
      responses:
        '204':
          description: deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...